refactor: 게이머 목록을 플레이어 목록으로 수정하라
